Fixed 상품정보 content 추가 및 id 값 int 에서 long으로 변경

+++ src/main/java/com/shop/controller/FileController.java
@@ -19,7 +19,7 @@ public class FileController {
     private FileService fileService;
 
     @GetMapping(value="/img/{id}")
-    public ResponseEntity<Resource> getImg(@PathVariable int id){
+    public ResponseEntity<Resource> getImg(@PathVariable Long id){
         return fileService.getImg(id);
     }
