+++ src/main/java/com/shop/controller/ProductController.java
@@ -6,7 +6,6 @@ import com.shop.domain.request.ProductInfo;
 import com.shop.response.ResponseData;
 import com.shop.service.ProductService;
 import lombok.RequiredArgsConstructor;
-import org.apache.ibatis.annotations.Delete;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -28,7 +27,7 @@ public class ProductController {
     }
 
     @GetMapping(value="/product/{id}")
-    public ResponseEntity<ResponseData> getProduct(@PathVariable int id){
+    public ResponseEntity<ResponseData> getProduct(@PathVariable Long id){
         return productService.getProduct(id);
     }
